requirements.txt
flask==2.3.3
flask-cors==5.0.1
google-generativeai==0.3.1
gunicorn==21.2.0
python-dotenv==1.0.0 